requirements.txt
numpy>=1.20.0
Pillow>=8.2.0
wheel
